requirements.txt
matplotlib==3.3.3
numpy==1.19.3
opencv_python==4.5.4.60
pycryptodome==3.15.0
PyQt5==5.15.7
